tools/nl/detection_evals/requirements.txt
numpy
python-dateutil
pydantic
scikit-learn
pandas
absl-py
typing-extensions
requests
Flask==2.3.2